requirements.txt
gradio
fastapi
uvicorn
passlib[bcrypt]
requests
faiss-cpu
openai
